Migrate LAB tests to .NET Core
