+++ source/Tests/Logging.TestSupport/Logging.TestSupport.csproj
@@ -2,7 +2,7 @@
   <Import Project="..\..\..\package.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net47</TargetFrameworks>
+    <TargetFrameworks>net47;netcoreapp3.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
 
     <AssemblyVersion>$(MajorVersion).0.0.0</AssemblyVersion>
@@ -41,6 +41,12 @@
     <Reference Include="System.Messaging" />
   </ItemGroup>
 
+  <ItemGroup Condition="$(TargetFramework.StartsWith('netstandard')) OR $(TargetFramework.StartsWith('netcoreapp'))">
+    <Compile Remove="**\*Msmq*.cs" />
+    <Compile Remove="**\*EventLog*.cs" />
+    <Compile Remove="**\*Desktop*.cs" />
+  </ItemGroup>
+  
   <ItemGroup>
     <None Include="..\..\Logging.licenseheader">
       <Link>Logging.licenseheader</Link>
