+++ source/Tests/Logging/Formatters/XmlLogFormatterFixture.cs
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
 
-using System.Management.Instrumentation;
 using System.Xml;
 using Microsoft.Practices.EnterpriseLibrary.Logging.Tests;
 using Microsoft.Practices.EnterpriseLibrary.Logging.TestSupport;
@@ -97,7 +96,6 @@ namespace Microsoft.Practices.EnterpriseLibrary.Logging.Formatters.Tests
     // (because they inherit from LogEntry, an InstrumentationClass
     public class LogEntryWithIndexer : LogEntry
     {
-        [IgnoreMember]
         public string this[int index]
         {
             get { return null; }
@@ -106,7 +104,6 @@ namespace Microsoft.Practices.EnterpriseLibrary.Logging.Formatters.Tests
 
     public class LogEntryWithWriteOnlyProperty : LogEntry
     {
-        [IgnoreMember]
         public string WriteOnly
         {
             set { ; }
