+++ source/Tests/Logging/NetCoreHelper.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.Text;
+
+namespace Microsoft.Practices.EnterpriseLibrary.Logging.Tests
+{
+    public class NetCoreHelper
+    {
+        public static ConfigurationSection LookupConfigSection(string sectionName)
+        {
+            var configMap = new ExeConfigurationFileMap();
+            configMap.ExeConfigFilename = "Microsoft.Practices.EnterpriseLibrary.Logging.Tests.NetCore.dll.config";
+            var config = ConfigurationManager.OpenMappedExeConfiguration(configMap, ConfigurationUserLevel.None);
+            return config.GetSection(sectionName);
+        }
+    }
+}
