+++ source/Tests/Logging/TraceListeners/TestCustomObject.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Practices.EnterpriseLibrary.Logging.TraceListeners.Tests
+{
+    public class TestCustomObject
+    {
+        public override string ToString()
+        {
+            return "TestCustomObject";
+        }
+    }
+}
